Reorganise the repo to group dto and entities together
